+++ buf.h
@@ -7,18 +7,18 @@ enum escape { ESCAPE_NONE, ESCAPE_QUESTION, ESCAPE_C };
 
 typedef struct buf {
     char *data;
-    int size;
-    int pos;
-    int screenpos;
+    size_t size;
+    size_t pos;
+    size_t screenpos;
 } Buf;
 
 Buf *newbuf(void);
 void freebuf(Buf *buf);
-void bufappend(Buf *buf, char *string, int width, bool printable);
+void bufappend(Buf *buf, char *string, size_t width, bool printable);
 void bufappendchar(Buf *buf, char c);
 char *bufstring(Buf *buf);
-int bufpos(Buf *buf);
-int bufscreenpos(Buf *buf);
+size_t bufpos(Buf *buf);
+size_t bufscreenpos(Buf *buf);
 void printtobuf(const char *text, enum escape escape, Buf *buf);
 
 #endif
